+++ domain/user/src/main/kotlin/io/github/siyual_park/user/domain/auth/UserPrincipalClaimEmbeddingProcessor.kt
@@ -1,4 +1,4 @@
-package io.github.siyual_park.user.domain
+package io.github.siyual_park.user.domain.auth
 
 import io.github.siyual_park.auth.domain.token.ClaimEmbeddingProcessor
 import io.github.siyual_park.auth.domain.token.Claims
