+++ domain/user/src/main/kotlin/io/github/siyual_park/user/domain/auth/UserPrincipalRefreshProcessor.kt
@@ -1,6 +1,7 @@
-package io.github.siyual_park.user.domain
+package io.github.siyual_park.user.domain.auth
 
 import io.github.siyual_park.auth.domain.principal_refresher.PrincipalRefreshProcessor
+import io.github.siyual_park.user.domain.UserScopeFinder
 import kotlinx.coroutines.flow.toSet
 import org.springframework.stereotype.Component
